+++ MailslotReader.h
@@ -32,7 +32,7 @@ protected:
 	/// \id Unique id of the mailslot. Usually it is id of the current process.
 	/// If exception occurs, terminate() will be called internally.
 	template <size_t PREFIX_SIZE>
-	void initialize (const WCHAR (&prefix) [PREFIX_SIZE], DWORD id, size_t max_msg_size, CompletionPort& port)
+	void initialize (const WCHAR (&prefix) [PREFIX_SIZE], DWORD id, DWORD max_msg_size, CompletionPort& port)
 	{
 		WCHAR name [PREFIX_SIZE + 8];
 		_ultow (id, real_copy (prefix, prefix + PREFIX_SIZE - 1, name), 16);
